// chapurwidgets/src/app/models/pay.model.ts
export class Pay {
  amount: number;
  idStore: number;
  email: string;
  createDate: string;
  idPurchaseOrder: number;
  platformId: number;
  userPlatform: number;
  passwordPlatform: string;
  token: number;
  noCreditCard: string;
  nameCreditCard: string;
  months: number;
  withInterest: number;
  CVV: number;
}